refactor(dashboard-api): move project upsert behind the management service
The upsert grew past what belonged in a handler: three functions, a
type, three sentinels and a tier constant, none of them touching gin.
Keeping it in the web layer meant that layer importing dberrors and
authqueries, and knowing a Postgres constraint name to tell a taken slug
from a real failure.

The service now reports outcomes as sentinel errors and owns the cache
eviction its own write invalidates, the same rule the membership
operations already followed. The handler parses, calls, maps sentinels
to status codes and responds.

Splits the package by feature — service.go for the type, members.go and
project.go for the operations — with the tests following the same split.
